move increment* to own file extend ReferencableBase for IncrementImpl
